ramips: add support for Joowin WR758AC V1 and V2
This patch adds support for Joowin (aka Comfast) WR758AC V1 and V2 devices.

Both are MT7628AN based, but V1 has a MT7662 as its 5GHz chip, while V2 has a MT7663 as its 5GHz chip. Some units are apparently mislabelled as V1, so if when you boot OpenWRT you don't see the 5GHz WiFi chip in LuCI, try the other image.

To install, setup a TFTP server on a machine with IP address 192.168.1.10 connected to the device's ethernet port and rename the squashfs-sysupgrade.bin to firmware_auto.bin. Turn off the device using the physical switch, wait a few seconds and turn it on again. The device should pull the firmware_auto.bin from the TFTP server, when it does disconnect the cable or stop the TFTP server, and the device should boot up with OpenWRT installed and available on the usual 192.168.1.1 address after a while.

Signed-off-by: Rodrigo Araujo <araujo.rm@gmail.com>